+++ tinper-acs/ac-form-layout/src/index.js
@@ -1,8 +1,10 @@
 import FormLayout from './FormLayout';
 import FormItem from './FormItem';
 import FormRow from './FormRow';
+import FormItemSpan from './FormItemSpan';
 
 FormLayout.FormRow = FormRow;
 FormLayout.FormItem = FormItem;
+FormLayout.FormItemSpan = FormItemSpan;
 
 export default FormLayout;